+++ CHANGES.md
@@ -4,6 +4,8 @@ Unreleased
 1.7.3
 =====
 
+* [#255](https://github.com/serokell/universum/issues/255):
+  Add `groupByFst` and `groupByKey`.
 * [#236](https://github.com/serokell/universum/issues/236):
   Add `updateMVar'` and `updateTVar'`.
 * [#244](https://github.com/serokell/universum/issues/244)